fix: wrong return type for getDurationFieldValue
The numeric field can return a float value.

You should not have any errors on an artifact that is using a duration
field as a float. (The same in the roadmap widget).

Fix request #21401: Duration field can be a float but Roadmap do not take it into account

+++ plugins/tracker/include/Tracker/Semantic/Timeframe/TimeframeArtifactFieldsValueRetriever.php
@@ -56,7 +56,7 @@ class TimeframeArtifactFieldsValueRetriever
      * @throws TimeframeFieldNotFoundException
      * @throws TimeframeFieldNoValueException
      */
-    public static function getDurationFieldValue(\Tracker_FormElement_Field_Numeric $duration_field, \PFUser $user, Artifact $artifact): ?int
+    public static function getDurationFieldValue(\Tracker_FormElement_Field_Numeric $duration_field, \PFUser $user, Artifact $artifact): ?float
     {
         if (! $duration_field->userCanRead($user)) {
             throw new TimeframeFieldNotFoundException();
